[1.1.0] Full package refactoring Adapters -> Serializer for better readability README.md updated
