+++ src/main/java/io/goodforgod/gson/configuration/serializer/InstantSerializer.java
@@ -1,0 +1,38 @@
+package io.goodforgod.gson.configuration.serializer;
+
+import com.google.gson.*;
+import java.lang.reflect.Type;
+import java.time.Instant;
+import java.time.format.DateTimeFormatter;
+
+/**
+ * @see Instant
+ * @author Anton Kurako (GoodforGod)
+ * @since 25.04.2021
+ */
+public class InstantSerializer implements JsonSerializer<Instant>, JsonDeserializer<Instant> {
+
+    private final DateTimeFormatter formatter;
+
+    public InstantSerializer() {
+        this(DateTimeFormatter.ISO_INSTANT);
+    }
+
+    public InstantSerializer(DateTimeFormatter formatter) {
+        this.formatter = formatter;
+    }
+
+    @Override
+    public Instant deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
+        try {
+            return formatter.parse(json.getAsString()).query(Instant::from);
+        } catch (Exception e) {
+            throw new JsonParseException(e);
+        }
+    }
+
+    @Override
+    public JsonElement serialize(Instant src, Type typeOfSrc, JsonSerializationContext context) {
+        return new JsonPrimitive(formatter.toFormat(Instant::from).format(src));
+    }
+}
