--- src/main/java/io/gson/adapters/LocalDateTimeAdapter.java
@@ -1,38 +1,0 @@
-package io.gson.adapters;
-
-import com.google.gson.*;
-import java.lang.reflect.Type;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-/**
- * @see LocalDateTime
- * @author Anton Kurako (GoodforGod)
- * @since 25.04.2021
- */
-public class LocalDateTimeAdapter implements JsonSerializer<LocalDateTime>, JsonDeserializer<LocalDateTime> {
-
-    private final DateTimeFormatter formatter;
-
-    public LocalDateTimeAdapter() {
-        this(DateTimeFormatter.ISO_LOCAL_DATE_TIME);
-    }
-
-    public LocalDateTimeAdapter(DateTimeFormatter formatter) {
-        this.formatter = formatter;
-    }
-
-    @Override
-    public LocalDateTime deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
-        try {
-            return formatter.parse(json.getAsString()).query(LocalDateTime::from);
-        } catch (Exception e) {
-            throw new JsonParseException(e);
-        }
-    }
-
-    @Override
-    public JsonElement serialize(LocalDateTime src, Type typeOfSrc, JsonSerializationContext context) {
-        return new JsonPrimitive(formatter.toFormat(LocalDateTime::from).format(src));
-    }
-}
