--- src/main/java/io/gson/adapters/OffsetDateTimeAdapter.java
@@ -1,38 +1,0 @@
-package io.gson.adapters;
-
-import com.google.gson.*;
-import java.lang.reflect.Type;
-import java.time.OffsetDateTime;
-import java.time.format.DateTimeFormatter;
-
-/**
- * @see OffsetDateTime
- * @author Anton Kurako (GoodforGod)
- * @since 25.04.2021
- */
-public class OffsetDateTimeAdapter implements JsonSerializer<OffsetDateTime>, JsonDeserializer<OffsetDateTime> {
-
-    private final DateTimeFormatter formatter;
-
-    public OffsetDateTimeAdapter() {
-        this(DateTimeFormatter.ISO_OFFSET_DATE_TIME);
-    }
-
-    public OffsetDateTimeAdapter(DateTimeFormatter formatter) {
-        this.formatter = formatter;
-    }
-
-    @Override
-    public OffsetDateTime deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
-        try {
-            return formatter.parse(json.getAsString()).query(OffsetDateTime::from);
-        } catch (Exception e) {
-            throw new JsonParseException(e);
-        }
-    }
-
-    @Override
-    public JsonElement serialize(OffsetDateTime src, Type typeOfSrc, JsonSerializationContext context) {
-        return new JsonPrimitive(formatter.toFormat(OffsetDateTime::from).format(src));
-    }
-}
