--- src/main/java/io/gson/adapters/ZonedDateTimeAdapter.java
@@ -1,38 +1,0 @@
-package io.gson.adapters;
-
-import com.google.gson.*;
-import java.lang.reflect.Type;
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
-
-/**
- * @see ZonedDateTime
- * @author Anton Kurako (GoodforGod)
- * @since 25.04.2021
- */
-public class ZonedDateTimeAdapter implements JsonSerializer<ZonedDateTime>, JsonDeserializer<ZonedDateTime> {
-
-    private final DateTimeFormatter formatter;
-
-    public ZonedDateTimeAdapter() {
-        this(DateTimeFormatter.ISO_ZONED_DATE_TIME);
-    }
-
-    public ZonedDateTimeAdapter(DateTimeFormatter formatter) {
-        this.formatter = formatter;
-    }
-
-    @Override
-    public ZonedDateTime deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
-        try {
-            return formatter.parse(json.getAsString()).query(ZonedDateTime::from);
-        } catch (Exception e) {
-            throw new JsonParseException(e);
-        }
-    }
-
-    @Override
-    public JsonElement serialize(ZonedDateTime src, Type typeOfSrc, JsonSerializationContext context) {
-        return new JsonPrimitive(formatter.toFormat(ZonedDateTime::from).format(src));
-    }
-}
